Tests for parse_gff_attribute_string() and build_reference_tables() have been moved from unit_tests_counter2_0.py to unit_tests_hts_parsing.py to reflect their new location in hts_parsing.py
